requirements.txt
watchdog
pillow